virtual_ship package (#22)
- Packaged code in virtual_ship package
- Created pyproject.toml for project
- Made tiny changes to code that were still on Emma's computer and not on github
